CTAA-1382 - refactoring of packages
